requirements.txt
folium==0.19.5
geopy==2.4.1
gspread==6.2.0
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
protobuf==4.25.3
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
streamlit==1.37.1
streamlit_folium==0.24.1
streamlit_geolocation==0.0.10